requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
folium==0.15.0
python-dotenv==1.0.0
aiohttp==3.8.6
streamlit==1.25.0
streamlit-folium==0.13.0
plotly==5.15.0



